+++ modules/ar_agent_store.h
@@ -8,6 +8,7 @@
 typedef struct ar_agent_registry_s ar_agent_registry_t;
 typedef struct ar_agent_store_s ar_agent_store_t;
 typedef struct ar_methodology_s ar_methodology_t;
+typedef struct ar_log_s ar_log_t;
 
 /**
  * @file ar_agent_store.h
@@ -24,12 +25,13 @@ typedef struct ar_methodology_s ar_methodology_t;
 
 /**
  * Create a new agent store instance
+ * @param ref_log The log instance for error reporting (borrowed reference, can be NULL)
  * @param ref_registry The agent registry to work with (borrowed reference)
  * @param ref_methodology The methodology for method lookups (borrowed reference)
  * @return New agent store instance (ownership transferred), or NULL on failure
  * @note Ownership: Caller must destroy the returned agent store
  */
-ar_agent_store_t* ar_agent_store__create(ar_agent_registry_t *ref_registry, ar_methodology_t *ref_methodology);
+ar_agent_store_t* ar_agent_store__create(ar_log_t *ref_log, ar_agent_registry_t *ref_registry, ar_methodology_t *ref_methodology);
 
 /**
  * Destroy an agent store instance